updated straight Gauss path

+++ src/matrix.cpp
@@ -7,7 +7,7 @@ int main()
     size_t matrix_size = 0;
     std::cin >> matrix_size;
 
-    matrix::Matrix<double> matrix(matrix_size, matrix_size);
+    matrix::Matrix<long double> matrix(matrix_size, matrix_size);
 
     matrix::filling_matrix(matrix);
